Replace Diff Conflate CLI options with config options (#5205)
* Move differential conflation from a command line argument to a config option in DifferentialConflation.conf
* Update java services to use correct differential syntax
* Submodule hoot-ui 8d8760ed4..70f483d5d: remove unauthenticated git protocol
* fix service test
* Sonar findings

+++ hoot-core/src/main/cpp/hoot/core/algorithms/changeset/NodeDensityTaskGridGenerator.cpp
@@ -129,8 +129,8 @@ TaskGrid NodeDensityTaskGridGenerator::_calcNodeDensityTaskGrid(OsmMapPtr map)
   for (size_t idx = 0; idx < rawTaskGrid.size(); idx++)
   {
     TaskGrid::TaskGridCell taskGridCell;
-    taskGridCell.id = idx + 1;
-    taskGridCell.replacementNodeCount = nodeCounts[idx];
+    taskGridCell.id = static_cast<int>(idx + 1);
+    taskGridCell.replacementNodeCount = static_cast<int>(nodeCounts[idx]);
     taskGridCell.bounds = rawTaskGrid[idx];
     taskGrid.addCell(taskGridCell);
   }
